azure/quotas/requirements.txt
azure-functions==1.18.0
azure-identity==1.15.0
azure-keyvault-secrets==4.8.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
google-auth==2.28.2
google-auth-oauthlib==1.2.0
gspread==6.0.2
gspread-formatting==1.1.2
idna==3.6
isodate==0.6.1
msal==1.27.0
msal-extensions==1.1.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.1
portalocker==2.8.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
six==1.16.0
StrEnum==0.4.15
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
